Изменен Tabs без Bootstrap, изменено удаление в EmployeeController. Готово!

+++ OperationMonitoring/Controllers/EmployeeController.cs
@@ -8,6 +8,7 @@ using Microsoft.EntityFrameworkCore;
 using OperationMonitoring.ModelsIdentity.Security;
 using Microsoft.AspNetCore.DataProtection;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 
 namespace OperationMonitoring.Controllers
 {
@@ -17,9 +18,12 @@ namespace OperationMonitoring.Controllers
 
         private readonly ApplicationContext db;
         private readonly IDataProtector protector;
-        public EmployeeController(ApplicationContext db, IDataProtectionProvider dataProtectionProvider, DataProtectionPurposeStrings dataProtectionPurposeStrings)
+        private readonly UserManager<IdentityUser> userManager;
+        public EmployeeController(ApplicationContext db, IDataProtectionProvider dataProtectionProvider,
+            DataProtectionPurposeStrings dataProtectionPurposeStrings, UserManager<IdentityUser> userManager)
         {
             this.db = db;
+            this.userManager = userManager;
             protector = dataProtectionProvider.CreateProtector(dataProtectionPurposeStrings.EmployeeIdRouteValue);
         }
 
@@ -101,9 +105,20 @@ namespace OperationMonitoring.Controllers
         {
             try
             {
-                db.Employees.Remove(db.Employees.FirstOrDefault(x => x.Id == Convert.ToInt32(protector.Unprotect(id))));
-                db.SaveChanges();
-                return RedirectToAction("AdminPanel", "Admin");
+                var userId = userManager.GetUserId(HttpContext.User);
+                var employee = db.Employees.AsNoTracking().Include(i=>i.IdentityUser).FirstOrDefault(x => x.Id == Convert.ToInt32(protector.Unprotect(id)));
+               
+                if (employee.IdentityUser ==null || employee.IdentityUser.Id != userId)
+                {
+                    db.Employees.Remove(db.Employees.FirstOrDefault(x => x.Id == Convert.ToInt32(protector.Unprotect(id))));
+                    db.SaveChanges();
+                    return RedirectToAction("AdminPanel", "Admin");
+                }
+                else
+                {
+                    return Content("Извините, но вы не можете удалить свои данные! Для редактирование перейдите во вкладу List Employee!");
+                }
+                
             }
             catch { return View(); }
         }
